Refactod phase interferometer tests
